requirements.txt
albumentations==1.1.0
bert_score==0.3.11
datasets==2.3.2
evaluate==0.2.2
flask==3.0.2
imagesize==1.3.0
matplotlib==3.5.2
nltk==3.7
numpy==1.22.3
opencv_python==4.5.5.64
pandas==1.4.2
Pillow==9.2.0
pycocoevalcap==1.2
pycocotools==2.0.5
rouge-score==0.1.2
scikit_learn==1.1.2
setuptools==61.2.0
spacy>=3.0.0,<4.0.0
statsmodels==0.13.2
tensorboard==2.9.0
tensorboardX==2.5.1
torch==1.12.1
torchinfo==1.7.0
torchmetrics==0.9.3
torchvision==0.13.1
tqdm==4.64.0
transformers==4.19.2